\name{EasyUpliftTree-package}
\alias{EasyUpliftTree-package}
\alias{EasyUpliftTree}
\docType{package}
\title{
Easy Uplift Tree Model for R
}
\description{
Easy Uplift Tree Model for R
}
\details{
\tabular{ll}{
Package: \tab EasyUpliftTree\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-09-23\cr
License: \tab BSD\cr
}

}
\author{
Yohei Sato, Issei Kurahashi
Maintainer: Yohei Sato <yokkun@tkul.jp>
}
\references{
http://stochasticsolutions.com/sbut.html
}
\keyword{ package }
\seealso{

}
\examples{

}
