\name{print}
\alias{print}
\title{
    print
}
\description{
    print
}
\usage{
print(x,...)
}
\arguments{
  \item{x}{
    an object used to select a method
}
  \item{\dots}{
  print parameter.
}
}
\value{
None
}

\author{
Yohei Sato
}

\seealso{
  \code{\link{print.UpliftTreeNode}}
}

\examples{
\dontrun{
print(x,...)
}
}


