\name{print.UpliftTreeNode}
\alias{print.UpliftTreeNode}
\title{
    print for Uplift Tree
}
\description{
    print for Uplift Tree
}
\usage{
print.UpliftTreeNode(tree, indent = "")
}
\arguments{
  \item{tree}{
    UpliftTreeNode
}
  \item{indent}{
    indente chatacter.
}
}
\details{
}
\value{
None
}
\references{
}
\author{
Yohei Sato
}
\note{
}

\seealso{

}
\examples{
\dontrun{
print(tree)
}
}
\keyword{ ~kwd1 }

