\name{print.UpliftTreeNode}
\alias{print.UpliftTreeNode}
\title{
    print for Uplift Tree
}
\description{
    print for Uplift Tree
}
\usage{
print.UpliftTreeNode(tree, indent = "")
}
\arguments{
  \item{tree}{
    UpliftTreeNode
}
  \item{indent}{
    indente chatacter.
}
}

\value{
None
}

\author{
Yohei Sato
}

\examples{
\dontrun{
print(tree)
}
}


